Add RBAC to allow AAP job to publish templates to the fulfillment-service (#3)
* Fix pre-commit linting errors

There were existing linting errors in the repository that were causing
erroneous failures in this pull request.

* Add RBAC to allow AAP job to publish templates to the fulfillment-service

This adds the `template-publisher` service account to the `fulfillment-aap`
namespace, and grants it permission to create short-lived tokens for the
`controller` service account in the `innabox` namespace.

We will use these credentials to publish template information to the
fulfillment api.

* Relocate fulfillment-aap files

The cluster-scope directory is primarily for cluster-scoped resources:
namespaces, subscriptions (which often introduce new CRDs), etc. Namespace
level configuration should generally be located elsewhere.

This commit moves the fulfillment-aap related configuration into a
top-level `fulfillment-aap` directory; this will require adding a
corresponding `fulfillment-aap` application to argocd.

---------

Co-authored-by: Lars Kellogg-Stedman <lars@redhat.com>